+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Release Date
+
+## Release Version
+
+- [minor] Added support for file based initialization
+
 # 30-06-2025
 
 ## 2.02.1